plugins/diffusers/server/pyproject.toml
[project]
name = "viwo-diffusers-server"
version = "0.1.0"
description = "Stable Diffusion image generation server for viwo"
requires-python = ">=3.13,<3.14"
dependencies = [
    "diffusers (>=0.35.1,<0.36.0)",
    "torch (>=2.8.0,<3.0.0)",
    "transformers (>=4.56.1,<5.0.0)",
    "fastapi (>=0.115.0,<1.0.0)",
    "uvicorn[standard] (>=0.34.0,<1.0.0)",
    "pillow (>=11.0.0,<12.0.0)",
    "accelerate (>=1.2.1,<2.0.0)",
    "controlnet-aux (>=0.0.9,<1.0.0)",
    "opencv-python (>=4.10.0,<5.0.0)",
    "basicsr (>=1.4.2,<2.0.0)",
    "realesrgan (>=0.3.0,<1.0.0)",
    "gfpgan (>=1.3.8,<2.0.0)",
]

[tool.pyright]
typeCheckingMode = "strict"

[tool.ruff]
line-length = 100
target-version = "py313"
